+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New `bootstrap` option for `Waku.create` to easily connect to Waku nodes upon start up.
+
 ### Changed
-- Renamed `discover.getStatusFleetNodes` to `discovery.getBootstrapNodes` and make it more generic to allow retrieval of bootstrap nodes from other sources.
+- Renamed `discover.getStatusFleetNodes` to `discovery.getBootstrapNodes`; 
+  Changed the API to allow retrieval of bootstrap nodes from other sources.
 
 ### Removed
 - Examples (cli-chat): The focus of this library is Web environment;